# Add benchmark project and improve performance of constant array creation