--- a/ParquetSharp.Dataset.Benchmark/.gitignore
+++ b/ParquetSharp.Dataset.Benchmark/.gitignore
@@ -1,0 +1,1 @@
+BenchmarkDotNet.Artifacts