--- a/ParquetSharp.Dataset.Benchmark/Program.cs
+++ b/ParquetSharp.Dataset.Benchmark/Program.cs
@@ -1,0 +1,17 @@
+﻿using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Jobs;
+using BenchmarkDotNet.Running;
+
+namespace ParquetSharp.Dataset.Benchmark;
+
+internal static class Program
+{
+    public static void Main(string[] args)
+    {
+        var config = DefaultConfig.Instance
+            .WithOptions(ConfigOptions.Default | ConfigOptions.StopOnFirstError)
+            .AddJob(Job.ShortRun);
+
+        BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args, config);
+    }
+}