--- a/ParquetSharp.Dataset.sln
+++ b/ParquetSharp.Dataset.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParquetSharp.Dataset", "Par
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParquetSharp.Dataset.Test", "ParquetSharp.Dataset.Test\ParquetSharp.Dataset.Test.csproj", "{826AF241-4060-4D57-896A-19138A9C5DBB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParquetSharp.Dataset.Benchmark", "ParquetSharp.Dataset.Benchmark\ParquetSharp.Dataset.Benchmark.csproj", "{FFF22C4C-D90F-4753-A427-BB0BB49F5794}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{826AF241-4060-4D57-896A-19138A9C5DBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{826AF241-4060-4D57-896A-19138A9C5DBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{826AF241-4060-4D57-896A-19138A9C5DBB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FFF22C4C-D90F-4753-A427-BB0BB49F5794}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFF22C4C-D90F-4753-A427-BB0BB49F5794}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFF22C4C-D90F-4753-A427-BB0BB49F5794}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFF22C4C-D90F-4753-A427-BB0BB49F5794}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal