--- a/ParquetSharp.Dataset/InternalsVisibleTo.cs
+++ b/ParquetSharp.Dataset/InternalsVisibleTo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("ParquetSharp.Dataset.Test")]
+[assembly: InternalsVisibleTo("ParquetSharp.Dataset.Benchmark")]