This PR adds a `ParquetSharp.Dataset.Benchmark` project that uses BenchmarkDotNet. I've added some high-level benchmarks of dataset reading, and then micro benchmarks for creating constant valued arrays and applying filters.

Creating the constant columns for fields from the partitioning schema was taking a particularly long amount of time, causing the "ReadAllData" benchmark to be much slower than reading all files with ParquetSharp directly (the "ReadAllFilesDirectly") benchmark, so I've sped that up significantly.

I also improved the performance of filtering primitive arrays that don't have nulls present. This wasn't taking a significant amount of time in the high level benchmark but would be more important for filtering data with a large number of columns.

Benchmark results from my machine are:

## DatasetRead

Before:

| Method               | Mean      | Error     | StdDev    | Ratio | RatioSD |
|--------------------- |----------:|----------:|----------:|------:|--------:|
| ReadAllFilesDirectly | 21.133 ms |  2.266 ms | 0.1242 ms |  1.00 |    0.00 |
| ReadAllData          | 87.099 ms |  2.420 ms | 0.1326 ms |  4.12 |    0.03 |
| FilterPartitions     |  1.324 ms |  1.019 ms | 0.0558 ms |  0.06 |    0.00 |
| FilterFileData       | 49.531 ms | 17.072 ms | 0.9358 ms |  2.34 |    0.05 |
| FilterToFileColumns  | 27.018 ms |  4.716 ms | 0.2585 ms |  1.28 |    0.01 |
| FilterToSingleColumn | 16.426 ms |  1.302 ms | 0.0714 ms |  0.78 |    0.01 |

After:

| Method               | Mean        | Error        | StdDev    | Ratio | RatioSD |
|--------------------- |------------:|-------------:|----------:|------:|--------:|
| ReadAllFilesDirectly | 22,034.4 μs |  5,350.88 μs | 293.30 μs |  1.00 |    0.00 |
| ReadAllData          | 33,189.4 μs |  2,864.13 μs | 156.99 μs |  1.51 |    0.03 |
| FilterPartitions     |    785.8 μs |     39.23 μs |   2.15 μs |  0.04 |    0.00 |
| FilterFileData       | 47,252.1 μs | 10,079.04 μs | 552.47 μs |  2.14 |    0.01 |
| FilterToFileColumns  | 24,265.1 μs |  1,529.44 μs |  83.83 μs |  1.10 |    0.01 |
| FilterToSingleColumn | 16,486.8 μs | 15,736.79 μs | 862.59 μs |  0.75 |    0.05 |

## CreateConstantArray

Before:

| Method                    | NumRows | Mean       | Error       | StdDev    |
|-------------------------- |-------- |-----------:|------------:|----------:|
| **CreateConstantIntArray**    | **100**     |   **2.115 μs** |   **0.0935 μs** | **0.0051 μs** |
| CreateConstantStringArray | 100     |   5.121 μs |   0.5213 μs | 0.0286 μs |
| **CreateConstantIntArray**    | **1000**    |  **12.606 μs** |   **0.3827 μs** | **0.0210 μs** |
| CreateConstantStringArray | 1000    |  37.598 μs |   6.9776 μs | 0.3825 μs |
| **CreateConstantIntArray**    | **10000**   | **128.637 μs** |   **8.0865 μs** | **0.4432 μs** |
| CreateConstantStringArray | 10000   | 399.849 μs | 105.8396 μs | 5.8014 μs |

After:

| Method                    | NumRows | Mean        | Error     | StdDev   |
|-------------------------- |-------- |------------:|----------:|---------:|
| **CreateConstantIntArray**    | **100**     |  **1,221.9 ns** | **250.79 ns** | **13.75 ns** |
| CreateConstantStringArray | 100     |  2,712.4 ns | 739.45 ns | 40.53 ns |
| **CreateConstantIntArray**    | **1000**    |    **816.8 ns** |  **99.03 ns** |  **5.43 ns** |
| CreateConstantStringArray | 1000    |  5,612.4 ns | 421.04 ns | 23.08 ns |
| **CreateConstantIntArray**    | **10000**   |  **3,463.7 ns** |  **60.49 ns** |  **3.32 ns** |
| CreateConstantStringArray | 10000   | 35,203.2 ns | 882.23 ns | 48.36 ns |

## ApplyFilterMask

Before:

| Method                   | NumRows | Mean       | Error      | StdDev    |
|------------------------- |-------- |-----------:|-----------:|----------:|
| **FilterInt64Array**         | **100**     |   **2.625 μs** |  **0.2909 μs** | **0.0159 μs** |
| FilterNullableInt64Array | 100     |   3.230 μs |  0.2394 μs | 0.0131 μs |
| **FilterInt64Array**         | **1000**    |   **8.842 μs** |  **0.5678 μs** | **0.0311 μs** |
| FilterNullableInt64Array | 1000    |  15.823 μs |  3.9359 μs | 0.2157 μs |
| **FilterInt64Array**         | **10000**   | **100.862 μs** | **14.9931 μs** | **0.8218 μs** |
| FilterNullableInt64Array | 10000   | 180.643 μs |  1.1782 μs | 0.0646 μs |

After:

| Method                   | NumRows | Mean       | Error      | StdDev    |
|------------------------- |-------- |-----------:|-----------:|----------:|
| **FilterInt64Array**         | **100**     |   **1.365 μs** |  **0.2347 μs** | **0.0129 μs** |
| FilterNullableInt64Array | 100     |   2.953 μs |  5.3568 μs | 0.2936 μs |
| **FilterInt64Array**         | **1000**    |   **3.130 μs** |  **0.3341 μs** | **0.0183 μs** |
| FilterNullableInt64Array | 1000    |  15.758 μs |  1.2079 μs | 0.0662 μs |
| **FilterInt64Array**         | **10000**   |  **26.081 μs** |  **1.0576 μs** | **0.0580 μs** |
| FilterNullableInt64Array | 10000   | 188.546 μs | 27.3601 μs | 1.4997 μs |